Move UserLogToken into a User model concern
Review asked for this to live with the model it describes rather than in lib/.
Cdo::UserLogToken becomes User::LogToken at
dashboard/app/models/concerns/user/log_token.rb, and User includes it.

The concern adds three ways in, matching how callers actually hold a user:

  user.log_token(destination:)                      a loaded record
  User.log_token_for(user_id, destination:)         an id and no record
  User.resolve_log_token(token, actor_id:, reason:) the audited direction

log_token_for exists because the admin lookup page derives from an id it
deliberately does not load a row for.

The cipher stays on User::LogToken itself rather than moving into
class_methods, so User does not gain private class methods named keys, parse,
decrypt and audit. Behaviour is unchanged.

The test moves with it, from the lib suite to
dashboard/test/models/concerns/user/log_token_test.rb, and now runs under
Rails, which the lib suite could not do on a machine without a seeded
dashboard_test database.
